backend-ml-python/requirements.txt
# TwoInOne ML Backend - Dépendances Python

# Framework Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Reconnaissance Faciale et Vision
face-recognition==1.3.0
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# Base de données
psycopg2-binary==2.9.9

# Utilities
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.5.3

# Machine Learning (pour futures fonctionnalités)
scikit-learn==1.4.0
pandas==2.2.0

# Logging et monitoring
loguru==0.7.2
